Added from_inthandle() to the Entity and EntityDictionary classes.
